Move `ExtraMetadata` from `db::types` to `model`
