+++ backend/src/api/model/event.rs
@@ -24,10 +24,10 @@ use crate::{
         },
     },
     db::{
-        types::{EventCaption, EventSegment, EventState, EventTrack, ExtraMetadata, Credentials},
+        types::{EventCaption, EventSegment, EventState, EventTrack, Credentials},
         util::{impl_from_db, select},
     },
-    model::Key,
+    model::{Key, ExtraMetadata},
     prelude::*,
 };
