+++ backend/src/db/types.rs
@@ -112,57 +112,6 @@ pub enum TextAssetType {
 }
 
 
-/// Represents extra metadata in the DB. Is a map from "namespace" to a
-/// `string -> string array` map.
-///
-/// Each namespace key is a URL pointing to an XML namespace definition OR
-/// `"dcterms"` for the dc terms (most common namespace). The value for each
-/// namespace is a simple string-key map where each value is an array of string
-/// values.
-#[derive(Debug, Serialize, Deserialize, Default)]
-#[cfg_attr(test, derive(PartialEq, Eq))]
-pub(crate) struct ExtraMetadata {
-    /// Metadata of the dcterms
-    #[serde(default)]
-    pub(crate) dcterms: MetadataMap,
-
-    /// Extended metadata.
-    #[serde(flatten)]
-    pub(crate) extended: HashMap<String, MetadataMap>,
-}
-
-type MetadataMap = HashMap<String, Vec<String>>;
-
-impl ToSql for ExtraMetadata {
-    fn to_sql(
-        &self,
-        ty: &postgres_types::Type,
-        out: &mut BytesMut,
-    ) -> Result<postgres_types::IsNull, Box<dyn std::error::Error + Sync + Send>> {
-        serde_json::to_value(self)
-            .expect("failed to convert `ExtraMetadata` to JSON value")
-            .to_sql(ty, out)
-    }
-
-    fn accepts(ty: &postgres_types::Type) -> bool {
-        <serde_json::Value as ToSql>::accepts(ty)
-    }
-
-    postgres_types::to_sql_checked!();
-}
-
-impl<'a> FromSql<'a> for ExtraMetadata {
-    fn from_sql(
-        ty: &postgres_types::Type,
-        raw: &'a [u8],
-    ) -> Result<Self, Box<dyn std::error::Error + Sync + Send>> {
-        serde_json::from_value(<_>::from_sql(ty, raw)?).map_err(Into::into)
-    }
-
-    fn accepts(ty: &postgres_types::Type) -> bool {
-        <serde_json::Value as FromSql>::accepts(ty)
-    }
-}
 
 /// Represents the type for the `custom_action_roles` field from `32-custom-actions.sql`.
 /// This holds a mapping of actions to lists holding roles that are allowed
